requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
flask_restful
python-dotenv
requests
pytest
supabase
flask_cors
